Rename godot::bind -> godot::register
Originally intended for "binding", the new term is more expressive and doesn't clash with other meanings
of "bind" (notably Gd::bind(), Callable::bind()).
