+++ godot-macros/src/derive/derive_export.rs
@@ -65,9 +65,9 @@ pub fn derive_export(decl: Declaration) -> ParseResult<TokenStream2> {
 
     let out = quote! {
         #[allow(unused_parens)]
-        impl godot::bind::property::Export for #name {
-            fn default_export_info() -> godot::bind::property::PropertyHintInfo {
-                godot::bind::property::PropertyHintInfo {
+        impl godot::register::property::Export for #name {
+            fn default_export_info() -> godot::register::property::PropertyHintInfo {
+                godot::register::property::PropertyHintInfo {
                     hint: godot::engine::global::PropertyHint::PROPERTY_HINT_ENUM,
                     hint_string: godot::prelude::GString::from(#hint_string),
                 }
