+++ godot-macros/src/util/mod.rs
@@ -35,12 +35,12 @@ pub fn class_name_obj(class: &impl ToTokens) -> TokenStream {
 
 pub fn property_variant_type(property_type: &impl ToTokens) -> TokenStream {
     let property_type = property_type.to_token_stream();
-    quote! { <<<#property_type as ::godot::bind::property::Property>::Intermediate as ::godot::builtin::meta::GodotConvert>::Via as ::godot::builtin::meta::GodotType>::Ffi::variant_type() }
+    quote! { <<<#property_type as ::godot::register::property::Property>::Intermediate as ::godot::builtin::meta::GodotConvert>::Via as ::godot::builtin::meta::GodotType>::Ffi::variant_type() }
 }
 
 pub fn property_variant_class_name(property_type: &impl ToTokens) -> TokenStream {
     let property_type = property_type.to_token_stream();
-    quote! { <<<#property_type as ::godot::bind::property::Property>::Intermediate as ::godot::builtin::meta::GodotConvert>::Via as ::godot::builtin::meta::GodotType>::class_name() }
+    quote! { <<<#property_type as ::godot::register::property::Property>::Intermediate as ::godot::builtin::meta::GodotConvert>::Via as ::godot::builtin::meta::GodotType>::class_name() }
 }
 
 pub fn bail_fn<R, T>(msg: impl AsRef<str>, tokens: T) -> ParseResult<R>
