+++ itest/rust/src/benchmarks/mod.rs
@@ -9,7 +9,7 @@
 
 use std::hint::black_box;
 
-use godot::bind::GodotClass;
+use godot::register::GodotClass;
 use godot::builtin::inner::InnerRect2i;
 use godot::builtin::{GString, Rect2i, StringName, Vector2i};
 use godot::engine::{Node3D, Os, RefCounted};
