+++ itest/rust/src/engine_tests/save_load_test.rs
@@ -5,7 +5,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-use godot::bind::GodotClass;
+use godot::register::GodotClass;
 use godot::engine::{load, save, try_load, try_save};
 use godot::obj::NewGd;
