+++ itest/rust/src/object_tests/onready_test.rs
@@ -6,7 +6,7 @@
  */
 
 use crate::framework::{expect_panic, itest};
-use godot::bind::{godot_api, GodotClass};
+use godot::register::{godot_api, GodotClass};
 use godot::engine::notify::NodeNotification;
 use godot::engine::INode;
