+++ itest/rust/src/register_tests/derive_variant_test.rs
@@ -7,7 +7,7 @@
 
 use std::fmt::Debug;
 
-use godot::bind::{FromGodot, GodotConvert, ToGodot};
+use godot::register::{FromGodot, GodotConvert, ToGodot};
 use godot::builtin::meta::{FromGodot, ToGodot};
 use godot::builtin::{dict, varray, Variant};
